✨ feat: Update errorHandler & resHandler
#
